Created run, plate and well for pacbio.

+++ db/migrate/20190624065732_create_pacbio_libraries.rb
@@ -1,8 +1,8 @@
 class CreatePacbioLibraries < ActiveRecord::Migration[5.2]
   def change
     create_table :pacbio_libraries do |t|
-      t.decimal :volume
-      t.decimal :concentration
+      t.float :volume
+      t.float :concentration
       t.string :library_kit_barcode
       t.integer :fragment_size
       t.belongs_to :pacbio_tag, index: true
